grpc-server/requirements.txt
grpcio==1.39.0
fake_useragent==0.1.11
requests==2.25.0
beautifulsoup4==4.9.3
protobuf==3.17.3
